int to float

+++ src/Model/Shapes/Shape.java
@@ -13,15 +13,15 @@ public abstract class Shape {
         return posX;
     }
 
-    public void setPosX(int posX) {
+    public void setPosX(float posX) {
         this.posX = posX;
     }
 
-    public int getPosY() {
+    public float getPosY() {
         return posY;
     }
 
-    public void setPosY(int posY) {
+    public void setPosY(float posY) {
         this.posY = posY;
     }
     public abstract String getDescription();
